+++ unnamed_engine/src/input/mod.rs
@@ -1,0 +1,4 @@
+//! Input
+pub mod keyboard;
+pub mod mouse;
+pub mod input_manager;